requirements.txt
openai
requests
beautifulsoup4
python-dotenv
python-dateutil
markdown
plotly
tiktoken